Add user management endpoints and integrate users blueprint
- Introduced a new `users.py` file with endpoints to retrieve user information and public keys by username and UUID.
- Registered the `users_bp` blueprint in the application.
- Removed outdated user retrieval endpoints from `files.py` to streamline functionality.
- Enhanced error handling and logging for user key retrieval processes.
